Clean up rspec helpers, tweak output
