+++ .rspec
@@ -1,4 +1,2 @@
---format Fuubar
 --colour
 --color
---require spec_helper
