+++ spec/controllers/teachers/classroom_activities_controller_spec.rb
@@ -1,4 +1,3 @@
-require 'spec_helper'
 require 'rails_helper'
 
 describe Teachers::ClassroomActivitiesController, type: :controller do
