+++ spec/spec_helper.rb
@@ -1,182 +1,25 @@
 require 'simplecov'
 require 'simplecov-json'
-SimpleCov.formatters = SimpleCov::Formatter::MultiFormatter.new([
-  SimpleCov::Formatter::HTMLFormatter,
-  SimpleCov::Formatter::JSONFormatter,
-])
-SimpleCov.start do
-  track_files '{app,lib}/**/*.rb'
-  add_filter '/spec/'
-end
 
-require 'rspec/retry'
-require 'rspec-redis_helper'
-
-require 'factory_bot_rails'
-require 'fileutils'
-
-# This method can be called in any of our backend tests to generate mockdata
-# for use in our frontend tests. Simply call the method with the response object
-# as a param, and this will automagically save the mock data in a file that
-# matches the request path name.
-# Ex: '/api/v1/cool_path' will be saved as 'api_v1_cool_path.json'
-def save_mock_data(response, condition=nil)
-  # If there is no folder for holding our mock data, create it.
-  FileUtils.mkdir_p 'client/__mockdata__'
-  condition = "_#{condition.downcase.gsub(/[^0-9a-z_]/, '_')}" if condition
-  File.write("client/__mockdata__/#{response.request.fullpath.gsub('/', '_')[1..-1]}#{condition}.json", response.body)
-end
-
-# This method can be called in any of our tests to convert a hash into a format
-# that allows it to be compared to a hash we generate from SQL. For example,
-# SQL will return 'f' instead of false. It's kinda hacky and shouldn't be
-# trusted in most circumstances. But, it's sufficient for most of our cases.
-def sanitize_hash_array_for_comparison_with_sql(array_of_hashes)
-  array_of_hashes.map do |hash|
-    hash.map do |key, value|
-      # Convert true and false to 't' and 'f'
-      value = value.to_s.first if value.in? [true, false]
-      # Convert numeric values to strings
-      value = value.to_s if value.is_a? Numeric
-      # Convert Times to have the right level of fidelity, and strip trailing zeroes
-      value = value.strftime('%Y-%m-%d %T.%6N').gsub(/0*$/, '') if value.is_a? Time
-      [key.to_s, value]
-    end.to_h
-  end
-end
-
-# This method can be called in any of our tests to convert a hash into a format
-# that allows it to be compared to a hash we pull back from Redis. Please note
-# that this is should not be trusted to be correct. Rather, it should be added
-# to and amended as necessary.
-def sanitize_hash_array_for_comparison_with_redis(array_of_hashes)
-  array_of_hashes.map do |hash|
-    hash.map do |key, value|
-      # Convert numeric values to strings
-      value = value.to_s if value.is_a? Numeric
-      [key.to_s, value]
-    end.to_h
-  end
-end
-
-# This file was generated by the `rails generate rspec:install` command. Conventionally, all
-# specs live under a `spec` directory, which RSpec adds to the `$LOAD_PATH`.
-# The generated `.rspec` file contains `--require spec_helper` which will cause this
-# file to always be loaded, without a need to explicitly require it in any files.
-#
-# Given that it is always loaded, you are encouraged to keep this file as
-# light-weight as possible. Requiring heavyweight dependencies from this file
-# will add to the boot time of your test suite on EVERY test run, even for an
-# individual file that may not need all of that loaded. Instead, consider making
-# a separate helper file that requires the additional dependencies and performs
-# the additional setup, and require it from the spec files that actually need it.
-#
-# The `.rspec` file also contains a few flags that are not defaults but that
-# users commonly want.
-#
-# See http://rubydoc.info/gems/rspec-core/RSpec/Core/Configuration
 RSpec.configure do |config|
-  # rspec-expectations config goes here. You can use an alternate
-  # assertion/expectation library such as wrong or the stdlib/minitest
-  # assertions if you prefer.
-
-  config.include FactoryBot::Syntax::Methods
-  config.include RSpec::RedisHelper, redis: true
-
-  # clean the Redis database around each run
-  # @see https://www.relishapp.com/rspec/rspec-core/docs/hooks/around-hooks
-  config.around( :each, redis: true ) do |example|
-    with_clean_redis do
-      example.run
-    end
-  end
-
-  config.verbose_retry = true
-  config.display_try_failure_messages = true
-  # these last two lines are for rspec-retry -- https://github.com/NoRedInk/rspec-retry
-  # add it to inconsistently failing tests
-  config.profile_examples = 10
-  #this prints the 10 slowest tests
+  config.formatter = :progress
   config.expect_with :rspec do |expectations|
-    # This option will default to `true` in RSpec 4. It makes the `description`
-    # and `failure_message` of custom matchers include text for helper methods
-    # defined using `chain`, e.g.:
-    # be_bigger_than(2).and_smaller_than(4).description
-    #   # => "be bigger than 2 and smaller than 4"
-    # ...rather than:
-    #   # => "be bigger than 2"
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
   end
 
-  # rspec-mocks config goes here. You can use an alternate test double
-  # library (such as bogus or mocha) by changing the `mock_with` option here.
   config.mock_with :rspec do |mocks|
-    # Prevents you from mocking or stubbing a method that does not exist on
-    # a real object. This is generally recommended, and will default to
-    # `true` in RSpec 4.
     mocks.verify_partial_doubles = true
   end
+end
 
-  # for allowing tests involving caching
-  # http://rosskaff.com/blog/2011/12/toggle-rails-caching-in-rspec-suite.html#.VO4tHVOS1oG
-  config.around(:each, :caching) do |example|
-    caching = ActionController::Base.perform_caching
-    ActionController::Base.perform_caching = example.metadata[:caching]
-    example.run
-    ActionController::Base.perform_caching = caching
+if ENV['CONTINUOUS_INTEGRATION'] == true
+  SimpleCov.formatters = SimpleCov::Formatter::MultiFormatter.new([
+    SimpleCov::Formatter::HTMLFormatter,
+    SimpleCov::Formatter::JSONFormatter,
+  ])
+
+  SimpleCov.start do
+    track_files '{app,lib}/**/*.rb'
+    add_filter '/spec/'
   end
-
-  # Allow Faker to reuse unique values between tests.
-  config.around(:each) do |example|
-    Faker::UniqueGenerator.clear
-    example.run
-  end
-
-
-
-
-# The settings below are suggested to provide a good initial experience
-# with RSpec, but feel free to customize to your heart's content.
-=begin
-  # These two settings work together to allow you to limit a spec run
-  # to individual examples or groups you care about by tagging them with
-  # `:focus` metadata. When nothing is tagged with `:focus`, all examples
-  # get run.
-  config.filter_run :focus
-  config.run_all_when_everything_filtered = true
-
-  # Limits the available syntax to the non-monkey patched syntax that is recommended.
-  # For more details, see:
-  #   - http://myronmars.to/n/dev-blog/2012/06/rspecs-new-expectation-syntax
-  #   - http://teaisaweso.me/blog/2013/05/27/rspecs-new-message-expectation-syntax/
-  #   - http://myronmars.to/n/dev-blog/2014/05/notable-changes-in-rspec-3#new__config_option_to_disable_rspeccore_monkey_patching
-  config.disable_monkey_patching!
-
-  # Many RSpec users commonly either run the entire suite or an individual
-  # file, and it's useful to allow more verbose output when running an
-  # individual spec file.
-  if config.files_to_run.one?
-    # Use the documentation formatter for detailed output,
-    # unless a formatter has already been configured
-    # (e.g. via a command-line flag).
-    config.default_formatter = 'doc'
-  end
-
-  # Print the 10 slowest examples and example groups at the
-  # end of the spec run, to help surface which specs are running
-  # particularly slow.
-  config.profile_examples = 10
-
-  # Run specs in random order to surface order dependencies. If you find an
-  # order dependency and want to debug it, you can fix the order by providing
-  # the seed, which is printed after each run.
-  #     --seed 1234
-  config.order = :random
-
-  # Seed global randomization in this process using the `--seed` CLI option.
-  # Setting this allows you to use `--seed` to deterministically reproduce
-  # test failures related to randomization by passing the same `--seed` value
-  # as the one that triggered the failure.
-  Kernel.srand config.seed
-=end
 end
